+++ README.md
@@ -1,0 +1,20 @@
+# If Statement Lab
+
+## Pre-Requisites
+You are expected to review the `Flow Control Statements` written lecture prior to working on this lab. If you would like to read additional resources, we recommend: https://www.w3schools.com/java/java_conditions.asp
+
+## Description
+In this exercise, you will practice creating if statements.
+    
+## Completing the Lab
+You will be editing the file `IfStatements.java`. This file can be found under the `src/main/java` package. Review the file's code and code comments for further instructions.
+
+Tasks to complete within the `IfStatements.java` file:
+- [ ] Revise the  method named `ifExample` so that it returns the correct int value.
+
+## Running Tests
+If you look within the `src/test/java` folder, you will find the Java class `IfStatements.java`.
+
+To run the tests, click on the Testing tab (test beaker icon on the left-hand side) and run the test cases by using the "Run Test" button (play icon). Please ensure that you do not already have the program running.
+
+You can optionally run the program for manual testing by switching to the file named `Main.java` under the `src/main/java` package and clicking the "Run Java" button (play icon) on the top-right of VSCode.
