+++ src/main/java/IfStatements.java
@@ -1,0 +1,16 @@
+
+public class IfStatements {
+    /** 
+     * This method should return the parameter `x` if the parameter `bool` is true, otherwise return the parameter `y`.
+     * 
+     * Using the paremeters `bool`, `x`, and `y`, to return the correct value. 
+     *
+     * @param bool - a true/false value.
+     * @param x - first number to be compared.
+     * @param y - second number to be compared.
+     * @return - x if bool is true, otherwise return y.
+     */
+    public int ifExample(boolean bool, int x, int y){
+        return 0;
+    }
+}
